requirements.txt
bottle==0.12.19
cached-property==1.5.1
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
dataclasses==0.8
docker==2.6.1
docker-compose==1.18.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
greenlet==1.1.1
idna==2.10
jsonschema==2.5.1
msgpack==1.0.2
Pygments==2.10.0
pynvim==0.4.3
PySocks==1.6.8
python-dotenv==0.19.1
PyYAML==3.13
requests==2.14.2
rich==10.12.0
six==1.14.0
texttable==1.6.2
typing-extensions==3.10.0.2
urllib3==1.25.6
websocket-client==0.47.0
